remove code from __init__ files
